Split up CardSearchComponent into two components
Top one contains 'chrome' around the topBar, the second is just the
results list.

Working idea is to create a second 'chrome' for a SplitView controller
that has different button behavior
