+++ components/CardSearchComponent/index.js
@@ -2,29 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { forEach } from 'lodash';
 import {
-  Button,
   Keyboard,
-  StyleSheet,
-  Text,
-  View,
 } from 'react-native';
 import { connectRealm } from 'react-native-realm';
 import { Navigation } from 'react-native-navigation';
 
 import L from '../../app/i18n';
-import CardSearchBox from './CardSearchBox';
 import {
   SORT_BY_TYPE,
-  SORT_BY_ENCOUNTER_SET,
 } from '../CardSortDialog/constants';
-import CardResultList from './CardResultList';
-import Switch from '../core/Switch';
+import CardSearchResultsComponent from '../CardSearchResultsComponent';
 import { iconsMap } from '../../app/NavIcons';
-import { filterToQuery } from '../../lib/filters';
 import calculateDefaultFilterState from '../filter/DefaultFilterState';
-import { MYTHOS_CARDS_QUERY, PLAYER_CARDS_QUERY } from '../../data/query';
-import typography from '../../styles/typography';
-import { COLORS } from '../../styles/colors';
 
 class CardSearchComponent extends React.Component {
   static propTypes = {
@@ -49,27 +38,15 @@ class CardSearchComponent extends React.Component {
     super(props);
 
     this.state = {
-      headerVisible: true,
-      searchText: false,
-      searchFlavor: false,
-      searchBack: false,
-      searchTerm: '',
       selectedSort: props.sort || SORT_BY_TYPE,
       filters: props.defaultFilterState,
       mythosMode: false,
       visible: true,
     };
 
-    this._showHeader = this.showHeader.bind(this);
-    this._hideHeader = this.hideHeader.bind(this);
     this._cardPressed = this.cardPressed.bind(this);
     this._toggleMythosMode = this.toggleMythosMode.bind(this);
-    this._toggleSearchText = this.toggleSearchMode.bind(this, 'searchText');
-    this._toggleSearchFlavor = this.toggleSearchMode.bind(this, 'searchFlavor');
-    this._toggleSearchBack = this.toggleSearchMode.bind(this, 'searchBack');
     this._sortChanged = this.sortChanged.bind(this);
-    this._searchUpdated = this.searchUpdated.bind(this);
-    this._clearSearchTerm = this.searchUpdated.bind(this, '');
     this._setFilters = this.setFilters.bind(this);
     this._clearSearchFilters = this.clearSearchFilters.bind(this);
     this._showSearchFilters = this.showSearchFilters.bind(this);
@@ -119,26 +96,6 @@ class CardSearchComponent extends React.Component {
     this._navEventListener.remove();
   }
 
-  showHeader() {
-    if (!this.state.headerVisible) {
-      this.setState({
-        headerVisible: true,
-      });
-    }
-  }
-
-  hideHeader() {
-    const {
-      headerVisible,
-      searchTerm,
-    } = this.state;
-    if (headerVisible && searchTerm === '') {
-      this.setState({
-        headerVisible: false,
-      });
-    }
-  }
-
   cardPressed() {
     this.isOnTop = false;
   }
@@ -174,9 +131,9 @@ class CardSearchComponent extends React.Component {
   showSearchFilters() {
     const {
       componentId,
-      baseQuery,
       defaultFilterState,
       modal,
+      baseQuery,
     } = this.props;
     this.isOnTop = false;
     Navigation.push(componentId, {
@@ -212,8 +169,7 @@ class CardSearchComponent extends React.Component {
         passProps: {
           sortChanged: this._sortChanged,
           selectedSort: this.state.selectedSort,
-          query: this.query(),
-          searchTerm: this.state.searchTerm,
+          hasEncounterCards: this.state.mythosMode,
         },
         options: {
           layout: {
@@ -294,190 +250,6 @@ class CardSearchComponent extends React.Component {
     this.syncNavigationButtons(!mythosMode, filters);
   }
 
-  searchUpdated(text) {
-    this.setState({
-      searchTerm: text,
-    });
-  }
-
-  applyQueryFilter(query) {
-    const {
-      searchTerm,
-      searchText,
-      searchFlavor,
-      searchBack,
-    } = this.state;
-
-    if (searchTerm !== '') {
-      const parts = searchBack ? [
-        'name contains[c] $0',
-        'linked_card.name contains[c] $0',
-        'back_name contains[c] $0',
-        'linked_card.back_name contains[c] $0',
-        'subname contains[c] $0',
-        'linked_card.subname contains[c] $0',
-      ] : [
-        'renderName contains[c] $0',
-        'renderSubname contains[c] $0',
-      ];
-      if (searchText) {
-        parts.push('real_text contains[c] $0');
-        parts.push('linked_card.real_text contains[c] $0');
-        parts.push('traits contains[c] $0');
-        parts.push('linked_card.traits contains[c] $0');
-        if (searchBack) {
-          parts.push('back_text contains[c] $0');
-          parts.push('linked_card.back_text contains[c] $0');
-        }
-      }
-
-      if (searchFlavor) {
-        parts.push('flavor contains[c] $0');
-        parts.push('linked_card.flavor contains[c] $0');
-        if (searchBack) {
-          parts.push('back_flavor contains[c] $0');
-          parts.push('linked_card.back_flavor contains[c] $0');
-        }
-      }
-      query.push(`(${parts.join(' or ')})`);
-    }
-  }
-
-  filterQueryParts() {
-    const {
-      filters,
-    } = this.state;
-    return filterToQuery(filters);
-  }
-
-  query() {
-    const {
-      baseQuery,
-      mythosToggle,
-    } = this.props;
-    const {
-      selectedSort,
-      mythosMode,
-    } = this.state;
-    const queryParts = [];
-    if (mythosToggle) {
-      if (mythosMode) {
-        queryParts.push(MYTHOS_CARDS_QUERY);
-      } else {
-        queryParts.push(PLAYER_CARDS_QUERY);
-      }
-    }
-    if (baseQuery) {
-      queryParts.push(baseQuery);
-    }
-    queryParts.push('(altArtInvestigator != true)');
-    queryParts.push('(back_linked != true)');
-    this.applyQueryFilter(queryParts);
-    forEach(
-      this.filterQueryParts(),
-      clause => queryParts.push(clause));
-
-    if (selectedSort === SORT_BY_ENCOUNTER_SET) {
-      queryParts.push(`(encounter_code != null OR linked_card.encounter_code != null)`);
-    }
-    return queryParts.join(' and ');
-  }
-
-  renderHeader() {
-    const {
-      searchText,
-      searchFlavor,
-      searchBack,
-      searchTerm,
-    } = this.state;
-    return (
-      <CardSearchBox
-        value={searchTerm}
-        visible={this.state.headerVisible}
-        onChangeText={this._searchUpdated}
-        searchText={searchText}
-        searchFlavor={searchFlavor}
-        searchBack={searchBack}
-        toggleSearchText={this._toggleSearchText}
-        toggleSearchFlavor={this._toggleSearchFlavor}
-        toggleSearchBack={this._toggleSearchBack}
-      />
-    );
-  }
-
-  renderExpandModesButtons() {
-    const {
-      mythosToggle,
-    } = this.props;
-    const {
-      mythosMode,
-    } = this.state;
-    const hasFilters = this.filterQueryParts().length > 0;
-    if (!mythosToggle && !hasFilters) {
-      return null;
-    }
-    return (
-      <View>
-        { !!mythosToggle && (
-          <View style={styles.button}>
-            <Button
-              onPress={this._toggleMythosMode}
-              title={mythosMode ? L('Search Player Cards') : L('Search Encounter Cards')}
-            />
-          </View>
-        ) }
-        { !!hasFilters && (
-          <View style={styles.button}>
-            <Button
-              onPress={this._clearSearchFilters}
-              title={L('Clear Search Filters')}
-            />
-          </View>
-        ) }
-      </View>
-    );
-  }
-
-  renderExpandSearchButtons() {
-    const {
-      searchTerm,
-      searchText,
-      searchBack,
-    } = this.state;
-    if (!searchTerm) {
-      return this.renderExpandModesButtons();
-    }
-    return (
-      <View>
-        { searchTerm && (
-          <View style={styles.button}>
-            <Button
-              onPress={this._clearSearchTerm}
-              title={L('Clear "{{searchTerm}}" search', { searchTerm })}
-            />
-          </View>
-        ) }
-        { !searchText && (
-          <View style={styles.toggle}>
-            <Text style={[typography.text, styles.toggleText]}>
-              { L('Search Game Text') }
-            </Text>
-            <Switch value={false} onValueChange={this._toggleSearchText} />
-          </View>
-        ) }
-        { !searchBack && (
-          <View style={styles.toggle}>
-            <Text style={[typography.text, styles.toggleText]}>
-              { L('Search Card Backs') }
-            </Text>
-            <Switch value={false} onValueChange={this._toggleSearchBack} />
-          </View>
-        ) }
-        { this.renderExpandModesButtons() }
-      </View>
-    );
-  }
-
   render() {
     const {
       componentId,
@@ -487,39 +259,34 @@ class CardSearchComponent extends React.Component {
       limits,
       footer,
       showNonCollection,
+      mythosToggle,
+      baseQuery,
     } = this.props;
     const {
       selectedSort,
-      searchTerm,
       visible,
       mythosMode,
+      filters,
     } = this.state;
-    const query = this.query();
     return (
-      <View style={styles.wrapper}>
-        { this.renderHeader() }
-        <View style={styles.container}>
-          <CardResultList
-            componentId={componentId}
-            query={query}
-            searchTerm={searchTerm}
-            sort={selectedSort}
-            originalDeckSlots={originalDeckSlots}
-            deckCardCounts={deckCardCounts}
-            onDeckCountChange={onDeckCountChange}
-            limits={limits}
-            cardPressed={this._cardPressed}
-            showHeader={this._showHeader}
-            hideHeader={this._hideHeader}
-            expandSearchControls={this.renderExpandSearchButtons()}
-            visible={visible}
-            showNonCollection={mythosMode || showNonCollection}
-          />
-        </View>
-        { !!footer && <View style={[
-          styles.footer,
-        ]}>{ footer }</View> }
-      </View>
+      <CardSearchResultsComponent
+        componentId={componentId}
+        baseQuery={baseQuery}
+        mythosToggle={mythosToggle}
+        mythosMode={mythosMode}
+        sort={selectedSort}
+        showNonCollection={showNonCollection}
+        selectedSort={selectedSort}
+        filters={filters}
+        toggleMythosMode={this._toggleMythosMode}
+        clearSearchFilters={this._clearSearchFilters}
+        originalDeckSlots={originalDeckSlots}
+        deckCardCounts={deckCardCounts}
+        onDeckCountChange={onDeckCountChange}
+        limits={limits}
+        footer={footer}
+        visible={visible}
+      />
     );
   }
 }
@@ -534,40 +301,5 @@ export default connectRealm(CardSearchComponent, {
     return {
       defaultFilterState: calculateDefaultFilterState(cards),
     };
-  },
-});
-
-const styles = StyleSheet.create({
-  wrapper: {
-    position: 'relative',
-    backgroundColor: 'white',
-    flex: 1,
-  },
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    flexDirection: 'column',
-    justifyContent: 'flex-start',
-    borderTopWidth: 1,
-    borderColor: '#bbb',
-  },
-  footer: {
-    position: 'absolute',
-    left: 0,
-    bottom: 0,
-    width: '100%',
-    backgroundColor: 'red',
-  },
-  button: {
-    margin: 8,
-  },
-  toggle: {
-    margin: 8,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  toggleText: {
-    minWidth: '60%',
   },
 });
