+++ components/CardSearchResultsComponent/CardResultList.js
