+++ components/CardSearchResultsComponent/ShowNonCollectionFooter.js
